fix: double-precision Player physics + save-before-recenter + disk-load pool clear
- Player.java: switched position/velocity/spawnPoint from Vector3f to Vector3d
  to fix float precision loss at far coordinates (~16 units at 100M).
  Small velocity deltas (0.1-0.3) were swallowed by position += velocity.
  Added setPosition(), resetVelocity(), getPositionD(), getVelocityD().

- ChunkManager.java: moved chunk unload (save) before recenterIfNeeded()
  to prevent saveChunk() from reading a cleared indirection table, which
  caused all saves to produce count=0 (empty) files. Also added
  clearChunkPoolSlot() before disk-load to prevent stale voxel data
  from recycled pool slots leaking into newly loaded chunks.

- GameContext.java: updated dimension-switch spawn placement to use
  setPosition()/resetVelocity() instead of mutating getPosition() copy.
  Cached getPositionD() in adjustSpawnYAfterChunkLoad().

- CommandProcessor.java: /tp now parses coordinates as double instead
  of float to match new teleport(double,double,double) signature.

+++ src/main/java/com/voxel/game/CommandProcessor.java
@@ -80,9 +80,9 @@ public class CommandProcessor {
             return;
         }
         try {
-            float x = Float.parseFloat(parts[1]);
-            float y = Float.parseFloat(parts[2]);
-            float z = Float.parseFloat(parts[3]);
+            double x = Double.parseDouble(parts[1]);
+            double y = Double.parseDouble(parts[2]);
+            double z = Double.parseDouble(parts[3]);
             ctx.player.teleport(x, y, z);
             ctx.setStatus(String.format("Teleported to %.1f, %.1f, %.1f", x, y, z));
         } catch (NumberFormatException e) {
